Convert ReadInterrupt* to return int64_t time (#1417)
HAL_ReadInterruptRisingTimestamp and HAL_ReadInterruptFallingTimestamp
return time as a double.  Instead, keep the raw integer count and move the
double conversion into the C++ and Java code.  This enables comparison of the
time with other timers.

+++ hal/src/main/java/edu/wpi/first/hal/InterruptJNI.java
@@ -25,9 +25,9 @@ public class InterruptJNI extends JNIWrapper {
 
   public static native void disableInterrupts(int interruptHandle);
 
-  public static native double readInterruptRisingTimestamp(int interruptHandle);
+  public static native long readInterruptRisingTimestamp(int interruptHandle);
 
-  public static native double readInterruptFallingTimestamp(int interruptHandle);
+  public static native long readInterruptFallingTimestamp(int interruptHandle);
 
   public static native void requestInterrupts(int interruptHandle, int digitalSourceHandle,
                                               int analogTriggerType);
